Don't mix Rule things and Collection things in the same module.
