+++ src/Language/Tandem/Pragma.hs
@@ -1,6 +1,6 @@
 module Language.Tandem.Pragma where
 
-import Language.Tandem.Rule (Label)
+import Language.Tandem.Collection (Label)
 
 data Pragma = Comment String
             | BatchIO Label Label
