+++ src/Main.hs
@@ -4,6 +4,7 @@ import System.Environment
 import System.Exit
 import System.IO
 
+import qualified Language.Tandem.Collection as Collection
 import qualified Language.Tandem.Rule as Rule
 import qualified Language.Tandem.Parser as Parser
 import qualified Language.Tandem.Pragma as Pragma
@@ -18,17 +19,17 @@ main = do
             putStrLn $ show rule
         ["eval", fileName] -> do
             (_, rule) <- loadSource fileName
-            case Eval.rewrite rule Rule.emptyCollection of
+            case Eval.rewrite rule Collection.empty of
                 Just c -> do
-                    putStr $ Rule.depict c
+                    putStr $ Collection.depict c
                 Nothing -> do
                     return ()
         ["showeval", fileName] -> do
             (_, rule) <- loadSource fileName
-            putStrLn $ show $ Eval.rewrite rule Rule.emptyCollection
+            putStrLn $ show $ Eval.rewrite rule Collection.empty
         ["run", fileName] -> do
             (pragmas, rule) <- loadSource fileName
-            initState <- setUpPragmas pragmas Rule.emptyCollection
+            initState <- setUpPragmas pragmas Collection.empty
             case Eval.rewrite rule initState of
                 Just result -> tearDownPragmas pragmas result
                 Nothing -> exitWith $ ExitFailure 1
@@ -52,7 +53,7 @@ setUpPragmas [] state = do
 
 setUpPragmas (Pragma.BatchIO inputLabel outputLabel:rest) state = do
     input <- getContents
-    setUpPragmas rest $ Rule.put inputLabel input state
+    setUpPragmas rest $ Collection.put inputLabel input state
 
 setUpPragmas (_:rest) state = do
     setUpPragmas rest state
@@ -61,7 +62,7 @@ tearDownPragmas [] state = do
     return ()
 
 tearDownPragmas (Pragma.BatchIO inputLabel outputLabel:rest) state = do
-    putStrLn $ reverse $ Rule.get outputLabel state
+    putStrLn $ reverse $ Collection.get outputLabel state
     tearDownPragmas rest state
 
 tearDownPragmas (_:rest) state = do
